move metadata api to a separete file
